Avoid repeatedly cloning large output strings
Hold stdout and stderr in Arc refcounts so that the task can be cloned
or sent between threads without cloning the potentially-large
buffers.

In particular, because the logs were previously held inline in the
`State`, this makes manipulation of the BTree faster because the nodes
are (much) smaller.

Helps with https://github.com/Nukesor/pueue/issues/82.