refactor: Move group addition/removal to TaskHandler
Previously, the message handlers were responsible for manipulating the
State on group addition or removal.
As we now also have to touch the TaskHandler whenever a group is
added/removed, this logic has also been moved to the TaskHandler.

That way we can ensure at least some kind of atomic scope, where all
required datastructures will be changed according to the requested
operation.
This should prevent desynchronization between our datastructures to a
certain degree.