change: Remove group configuration from settings
Previously, the group configuration was located in the settings file.

However, this lead to some problems as groups can be added and removed
dynamically during runtime.
To sync these changes, we had to save the settings file each time
something changed. This isn't acceptable, as users expected the config
file to be immutable.

To solve this issue, the groups have been removed from the settings
and are now only saved in the state.