fix: Always use /tmp as cwd in tests
Previously, all tests used a temporary directory in $TEMPDIR as the cwd
for tasks. This allowed us to do isolate tests and exeuted binaries to
the max.

This however had to change with the new client tests.
These tests use templates and expected $TEMPDIR to always be /tmp.
Since this isn't the case, we have to opt for a common denominator
directory that should exist in all unix systems.

Since /tmp should still be there, I opted for this solution, as it
seemed to be the least intrusive one.