refactor: Revamp task state handling
Previously, the `Task` struct had lots of runtime related fields such as
- `enqueue_at`: When a task should be enqueued.
- `enqueued_at`: When a task has been enqueued.
- `start`: When a task was started.
- `end`: When a task finished.
- `result`: The outcome of a task.

The problem with these was that all of them only made sense when the
task was in a specific state.
E.g. `enqueue_at` was only necessary when a task was `Queued`,
`start` only made sense when a task was at least started, etc.

Whenever the state of a task changed, those invariants needed to be
enforced, which was prone to error as it was really easy to just forget
about something.

The new design moves all of those fields and moves them into the
`TaskStatus` enum as struct variants.

This made the code slightly more complex, but significantly more robust.